refactor(*): updated variant enum to avoid conflicts
