+++ apps/backoffice-v2/src/lib/blocks/variants/BlocksVariant/BlocksVariant.tsx
@@ -3,14 +3,14 @@ import { KybDemoBlocks } from '@/lib/blocks/variants/KybDemoBlocks/KybDemoBlocks
 import { DefaultBlocks } from '@/lib/blocks/variants/DefaultBlocks/DefaultBlocks';
 import { FunctionComponent } from 'react';
 import { TWorkflowById } from '@/domains/workflows/fetchers';
-import { WorkflowVariant } from '@ballerine/common';
+import { WorkflowDefinitionVariant } from '@ballerine/common';
 
 export const BlocksVariant: FunctionComponent<{
   workflowDefinition: Pick<TWorkflowById['workflowDefinition'], 'variant' | 'config' | 'version'>;
 }> = ({ workflowDefinition }) => {
   if (
     workflowDefinition?.version >= 0 &&
-    workflowDefinition?.variant === WorkflowVariant.KYB &&
+    workflowDefinition?.variant === WorkflowDefinitionVariant.KYB &&
     workflowDefinition?.config?.isDemo
   ) {
     return <KybDemoBlocks />;
@@ -18,7 +18,7 @@ export const BlocksVariant: FunctionComponent<{
 
   if (
     workflowDefinition?.version >= 0 &&
-    workflowDefinition?.variant === WorkflowVariant.MANUAL_REVIEW &&
+    workflowDefinition?.variant === WorkflowDefinitionVariant.MANUAL_REVIEW &&
     workflowDefinition?.config?.isLegacyReject
   ) {
     return <ManualReviewBlocks />;
