+++ packages/common/src/consts/index.ts
@@ -36,7 +36,7 @@ export const CommonWorkflowStates = {
   REVISION: 'revision',
 } as const;
 
-export const WorkflowVariant = {
+export const WorkflowDefinitionVariant = {
   MANUAL_REVIEW: 'MANUAL_REVIEW',
   KYB: 'KYB',
   DEFAULT: 'DEFAULT',
